platforms/lib/wx_sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "weixin-clawbot-python"
version = "0.1.0"
description = "Python SDK for Weixin Bot API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "er1cw00" },
]
keywords = ["weixin", "bot", "sdk", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.8.0",
    "pycryptodome>=3.15.0",
    "qrcode>=7.0.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-aiohttp>=1.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/er1cw00/weixin-clawbot-python"
Repository = "https://github.com/er1cw00/weixin-clawbot-python"
Documentation = "https://github.com/er1cw00/weixin-clawbot-python#readme"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
